- Added nullPointer protection in method clearDirectory
git-svn-id: http://gforge.hl7.org/svn/fhir/trunk@1070 2f0db536-2c49-4257-a3fa-e771ed206c19

+++ implementations/java/org.hl7.fhir.utilities/src/org/hl7/fhir/utilities/Utilities.java
@@ -185,13 +185,15 @@ public class Utilities {
   }
 
   public static void clearDirectory(String folder) {
-    String[] files = new CSFile(folder).list();
-    for (String f : files) {
-      File fh = new CSFile(folder+File.separatorChar+f);
-      if (fh.isDirectory()) 
-        clearDirectory(fh.getAbsolutePath());
-      fh.delete();
-    }
+	  String[] files = new CSFile(folder).list();
+	  if (files != null) {
+		  for (String f : files) {
+			  File fh = new CSFile(folder+File.separatorChar+f);
+			  if (fh.isDirectory()) 
+				  clearDirectory(fh.getAbsolutePath());
+			  fh.delete();
+		  }
+	  }
   }
 
   public static void createDirectory(String path) {
